tracing: mv UniformFields -> emfields.uniform
